Add a Projects page linking to our GitHub repos
Add a Projects page which uses Jekyll's ability to read custom data from
JSON and YAML files[1] to present a list of our GitHub repositories
based on JSON data pulled from the GitHub API.

The JSON data is stored in the `_data` directory and can be updated
using the Rake task added in this commit:

```bash
bundle exec rake update_repo_data
```

Note that we use the `.yml` extension rather than `.json` because, as
of version 2.0.3, Jekyll doesn't support data files that use the .json
extension[2], even though it will interpret them if they are given a
YAML extension (presumably because YAML is a superset of JSON). We can
use `.json` once [this
issue](https://github.com/jekyll/jekyll/pull/2369) is included in a
Jekyll release.

[1]: http://jekyllrb.com/docs/datafiles/
[2]: https://github.com/jekyll/jekyll/issues/2441