Exclude index page from navigation
Because we don't explicitly set a page title in index.md in YAML
front-matter[1], Jekyll was creating an empty `<a>` anchor tag linking
to `/`.

This was causing rendering issues once the site was uploaded to GitHub
Pages because that `<a>` tag was being rendered last:

```html
<div class="trigger">
          <a class="page-link" href="/guidelines/">Guidelines</a>
          <a class="page-link" href="/projects/">Projects</a>
          <a class="page-link" href="/"></a>
      </div>
```

In Chrome for Mac, the empty anchor was being rendered on a new line
which caused the header width to increase.

I could have solved this by adding a title to the index page (which did
fix the issue in my tests), but the title I wanted to use would be
unwieldy for a navigation link.

[1]: http://jekyllrb.com/docs/frontmatter/